ncurses 5.9 - patch 20140927
+ implement curs_set in win_driver.c
+ implement flash in win_driver.c
+ fix an infinite loop in win_driver.c if the command-window loses
  focus.
+ improve the non-buffered mode, i.e., NCURSES_CONSOLE2, of
  win_driver.c by temporarily changing the buffer-size to match the
  window-size to eliminate the scrollback.  Also enforce a minimum
  screen-size of 24x80 in the non-buffered mode.
+ modify generated misc/Makefile to suppress install.data from the
  dependencies if the --disable-db-install option is used, compensating
  for the top-level makefile changes used to add ncurses*-config in the
  20140920 changes (report by Steven Honeyman).

+++ test/dots_termcap.c
@@ -29,7 +29,7 @@
 /*
  * Author: Thomas E. Dickey
  *
- * $Id: dots_termcap.c,v 1.7 2013/09/28 21:50:35 tom Exp $
+ * $Id: dots_termcap.c,v 1.8 2014/09/25 09:00:56 tom Exp $
  *
  * A simple demo of the termcap interface.
  */
@@ -154,7 +154,7 @@ static void
 my_napms(int ms)
 {
 #if defined(__MINGW32__) || !HAVE_GETTIMEOFDAY
-    Sleep(ms);
+    Sleep((DWORD) ms);
 #else
     struct timeval data;
     data.tv_sec = 0;
